requirements.txt
aioboto3==11.2.0
aiohttp==3.10.11
aiofiles==24.1.0
botocore==1.29.76
fastapi==0.115.0
pydantic==2.4.0
uvicorn==0.23.2
tortoise-orm==0.20.0
python-multipart==0.0.20
python-telegram-bot==21.4
httpx==0.27.0
